Arreglando errores

+++ src/main/java/ar/edu/unju/fi/tp6/repository/ICompraDB.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import ar.edu.unju.fi.tp6.model.Compra;
 
-public interface ICompraDB extends JpaRepository<Compra, Integer>{
+public interface ICompraDB extends JpaRepository<Compra, Long>{
 
 }
